Split widget: panel state cards
